Outfit Planner
With this program, you can easily save and select your combinations. First you need to log in to the system. You click on the "Log in" section. Then you register to the system. After entering the system, you go to the "Add Clothes" section and add the pieces you want. You create your combination with "Select Clothes for Outfit" and you save the combination you created with the "Save the Created Outfit" section and exit.